+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Hello Gateway: Support cert-manager and ExternalDNS to be HelmReleases.
 
+
 ## [7.2.2] - 2026-06-18
 
 ### Changed